LayoutTests/imported/w3c/web-platform-tests/2dcontext/fill-and-stroke-styles/2d.gradient.radial.negative-expected.txt

2d.gradient.radial.negative
createRadialGradient() throws INDEX_SIZE_ERR if either radius is negative
Actual output:

PASS createRadialGradient() throws INDEX_SIZE_ERR if either radius is negative 
